bm-agent: sync + service path fixes (ready for PR)
