--- camera_software/bm_agent/bm-agent.service
@@ -1,22 +1,0 @@
-# /etc/systemd/system/bm-agent.service
-[Unit]
-Description=Bristlemouth Agent (RTC/CLOCK sync)
-After=network-online.target
-Wants=network-online.target
-
-[Service]
-Type=simple
-WorkingDirectory=/home/pi/bm_camera/camera_software/bm_agent
-ExecStart=/usr/bin/python3 -u /home/pi/bm_camera/camera_software/bm_agent/run_agent.py
-Restart=always
-RestartSec=2
-
-# simplest: run as root to allow time setting
-User=root
-# (If you prefer 'pi', replace with User=pi and add AmbientCapabilities=CAP_SYS_TIME)
-
-StandardOutput=journal
-StandardError=journal
-
-[Install]
-WantedBy=multi-user.target
