+++ camera_software/bm_agent/bm_agent/bus.py
@@ -1,116 +1,100 @@
-# bm_agent/bm_agent/bus.py
-from __future__ import annotations
-import os
-import sys
-import time
 from pathlib import Path
-from typing import Callable, Iterable, Any
+import sys
+# add the folder that contains your proven bm_serial.py
+sys.path.append(str(Path.home() / "bm_camera/camera_software/dev_rtc_reader"))
+from bm_serial import BristlemouthSerial
 
-# --- import your known-good BristlemouthSerial helper once, cleanly ---
-DEV_HELPER = Path.home() / "bm_camera/camera_software/dev_rtc_reader"
-if str(DEV_HELPER) not in sys.path:
-	sys.path.append(str(DEV_HELPER))
+import sys, time, binascii
 
-from bm_serial import BristlemouthSerial  # provides .uart, .bristlemouth_sub(), .bristlemouth_process()
+# import your proven UART implementation from the original folder
+sys.path.append(str((__file__).__class__(__file__)))  # no-op keeps mypy quiet
+# Explicit add of the original folder:
+import sys as _sys, pathlib as _pl
+_sys.path.append(str(_pl.Path.home() / "bm_camera/camera_software/dev_rtc_reader"))
 
-try:
-	import serial  # for type/timeout controls and exceptions
-except Exception:
-	serial = None  # still works; we just won't tweak timeouts explicitly
+from bm_serial import BristlemouthSerial
 
+# def _uart_safety(uart):
+# 	try:
+# 		uart.timeout = 0.1
+# 		uart.write_timeout = 0.5
+# 		if hasattr(uart, "rtscts"):  uart.rtscts = False
+# 		if hasattr(uart, "dsrdtr"):  uart.dsrdtr = False
+# 		if hasattr(uart, "xonxoff"): uart.xonxoff = False
+# 	except Exception:
+# 		pass
 
-# --- small UART hygiene helpers ---
-
-def _uart_safety(uart) -> None:
-	"""Set sane defaults: bounded timeouts, no flow control."""
+def _uart_safety(uart):
 	try:
-		# a little time for reads; and a generous write timeout so we don't insta-fail
-		if getattr(uart, "timeout", None) in (None, 0):
-			uart.timeout = 0.2
-		if hasattr(uart, "write_timeout"):
-			uart.write_timeout = 3.0  # was 0.5; lengthen to ride out slow peer startup
-
-		# disable HW/SW flow control unless you KNOW you need them
-		if hasattr(uart, "rtscts"):
-			uart.rtscts = False
-		if hasattr(uart, "dsrdtr"):
-			uart.dsrdtr = False
-		if hasattr(uart, "xonxoff"):
-			uart.xonxoff = False
+		uart.timeout = 0.1
+		uart.write_timeout = 0.5
+		if hasattr(uart, "rtscts"):  uart.rtscts = False
+		if hasattr(uart, "dsrdtr"):  uart.dsrdtr = False
+		if hasattr(uart, "xonxoff"): uart.xonxoff = False
 	except Exception:
 		pass
-
-def _settle(uart) -> None:
-	"""Clear buffers and give the peer a breath."""
+	
+# def open_bus(uart_device="/dev/serial0", baudrate=115200):
+# 	bm = BristlemouthSerial()  # your class uses /dev/serial0@115200 already
+# 	uart = getattr(bm, "uart", None)
+# 	if uart:
+# 		_uart_safety(uart)
+# 		print(f"[BUS] open on {uart.port}")
+# 	else:
+# 		print("[BUS][WARN] no uart found on BristlemouthSerial")
+# 	return bm
+def open_bus(uart_device="/dev/serial0", baudrate=115200):
 	try:
-		uart.reset_input_buffer()
-	except Exception:
-		pass
-	try:
-		uart.reset_output_buffer()
-	except Exception:
-		pass
-	time.sleep(0.05)
-
-
-# --- public API used by run_agent.py ---
-
-def open_bus(uart_device: str = "/dev/serial0", baudrate: int = 115200) -> BristlemouthSerial:
-	"""
-	Create your BristlemouthSerial wrapper. Your helper typically opens /dev/serial0 internally.
-	We don't fight that—just configure and log clearly.
-	"""
-	# Most versions of your helper take no args and open /dev/serial0@115200 by default.
-	bm = BristlemouthSerial()
-
+		bm = BristlemouthSerial()  # opens /dev/serial0 exclusively
+	except serial.SerialException as e:
+		msg = str(e)
+		if "exclusively lock port" in msg or "Resource temporarily unavailable" in msg:
+			print("[BUS][ERROR] /dev/serial0 is busy (owned by another process).")
+			print("  Hints:")
+			print("   • stop any services first: sudo systemctl stop bm-agent.service bm-rtc-listener.service")
+			print("   • see owner:               sudo lsof /dev/serial0")
+			print("   • then try again in the foreground")
+			sys.exit(2)
+		raise
+	
 	uart = getattr(bm, "uart", None)
 	if uart:
 		_uart_safety(uart)
-		real = os.path.realpath(uart.port)
-		print(f"[BUS] open port={uart.port} (real={real}) baud={getattr(uart, 'baudrate', baudrate)}")
-		_settle(uart)
+		print(f"[BUS] open on {uart.port}")
 	else:
-		print("[BUS][WARN] BristlemouthSerial.uart missing")
-
+		print("[BUS][WARN] no uart found on BristlemouthSerial")
 	return bm
 
-
-def subscribe_one(bm: BristlemouthSerial, topic: str,
-				  callback: Callable[[int, str, bytes], Any]) -> None:
-	"""Subscribe to exactly one topic (back-to-basics)."""
-	if not topic:
-		return
-	print(f"[SUB(req)] {topic!r}")
-	# NOTE: bm_serial may also print its own "[SUB]" line; that's expected.
-	bm.bristlemouth_sub(topic, callback)
-
-
-def subscribe_many(bm: BristlemouthSerial, topics: Iterable[str],
-				   callback: Callable[[int, str, bytes], Any]) -> None:
-	"""
-	Retained for compatibility, but just fans out to subscribe_one().
-	Not the cause of your timeouts; the actual UART write happens inside bristlemouth_sub().
-	"""
+def subscribe_many(bm, topics, callback):
 	for t in topics:
-		if t:
-			subscribe_one(bm, t, callback)
+		if not t: continue
+		print(f"[SUB] subscribing to '{t}'")
+		bm.bristlemouth_sub(t, callback)
 
-
-def loop(bm: BristlemouthSerial, should_stop) -> None:
-	"""
-	Drive the Bristlemouth reader. Always closes the port before returning.
-	"""
+# def loop(bm, should_stop):
+# 	try:
+# 		while not should_stop():
+# 			had = bm.bristlemouth_process(0.1)
+# 			if not had:
+# 				print(".", end="", flush=True)
+# 	finally:
+# 		try:
+# 			if bm.uart and bm.uart.is_open:
+# 				bm.uart.flush(); bm.uart.close()
+# 		except Exception:
+# 			pass
+# 		print("\n[BUS] closed")
+def loop(bm, should_stop):
 	try:
 		while not should_stop():
-			bm.bristlemouth_process(0.1)
+			had = bm.bristlemouth_process(0.1)
+			if not had:
+				print(".", end="", flush=True)
 	finally:
 		try:
-			print("[BUS] closing…")
-			if getattr(bm, "uart", None) and bm.uart.is_open:
-				try:
-					bm.uart.flush()
-				except Exception:
-					pass
+			if bm.uart and bm.uart.is_open:
+				bm.uart.flush()
 				bm.uart.close()
-		finally:
-			print("[BUS] closed")
+		except Exception:
+			pass
+		print("\n[BUS] closed")
