+++ camera_software/bm_agent/bm_agent/handlers/rtc.py
@@ -1,26 +1,20 @@
-# bm_agent/handlers/rtc.py
-from __future__ import annotations
 import struct
 from datetime import datetime, timezone
 
-def _plausible_epoch_us(v: int) -> bool:
-	return 946684800_000000 <= v <= 4102444800_000000  # 2000..2100
+def _plausible_epoch_us(val: int) -> bool:
+	return 946684800_000000 <= val <= 4102444800_000000  # 2000..2100
 
-def decode_epoch_dt_from_payload(data: bytes) -> datetime | None:
-	# Works for 8B and longer stamped payloads; prefer the newest 8B (tail) if available.
-	if len(data) >= 16:
-		(tail_us,) = struct.unpack("<Q", data[-8:])
-		if _plausible_epoch_us(tail_us):
-			return datetime.fromtimestamp(tail_us / 1e6, tz=timezone.utc)
+def decode_epoch_dt_from_payload(data: bytes):
+	# Works for short (8B) and stamped (>=8B) payloads
 	if len(data) >= 8:
-		(head_us,) = struct.unpack("<Q", data[:8])
-		if _plausible_epoch_us(head_us):
-			return datetime.fromtimestamp(head_us / 1e6, tz=timezone.utc)
+		(ts_us,) = struct.unpack("<Q", data[:8])
+		if _plausible_epoch_us(ts_us):
+			return datetime.fromtimestamp(ts_us / 1e6, tz=timezone.utc)
 	return None
 
-def handle(node_id: int, topic: str, data: bytes, ctx: dict) -> None:
+def handle(node_id, topic: str, data: bytes, ctx):
 	dt = decode_epoch_dt_from_payload(data)
 	if dt:
-		print(f"[RTC] node={node_id} topic={topic} {dt.isoformat()}")
+		print(f"[RTC] {dt.isoformat()} (µs since epoch)")
 	else:
-		print(f"[RTC] node={node_id} topic={topic} unknown payload len={len(data)}")
+		print(f"[RTC] unknown payload: 0x{data.hex()}")
