+++ camera_software/tools/setup_bm_camera.sh
